✨ Feat: 임시 주문 세션에 저장

+++ src/main/java/com/telegro/telegro/domain/cart/dto/response/CartListDTO.java
@@ -2,6 +2,7 @@ package com.telegro.telegro.domain.cart.dto.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -9,7 +10,7 @@ public record CartListDTO(
         boolean isLast,
         int totalPage,
         long totalElement,
-        double totalPrice,
+        BigDecimal totalPrice,
         List<CartResponseDTO> carts
 ) {
 }
